# src/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Multi-Tenant service",
        "title": "Multi-Tenant API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5002",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "this endpoint is used to log a user in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "login request body",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user logged in successfully",
                        "schema": {
                            "$ref": "#/definitions/model.GenericResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "this endpoint signs up a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "user signup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e. Pass in the tenant access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "signup request body",
                        "name": "signupRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.signupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "signup successful",
                        "schema": {
                            "$ref": "#/definitions/model.GenericResponse"
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "patch": {
                "description": "this endpoint is used to update any of the users record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "updateUserByID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "update user request body",
                        "name": "updateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.updateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "user updated successfully",
                        "schema": {
                            "$ref": "#/definitions/model.GenericResponse"
                        }
                    }
                }
            }
        },
        "/auth/user/{id}": {
            "get": {
                "description": "this endpoint gets a user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "getUserByID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user details fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/model.GenericResponse"
                        }
                    }
                }
            }
        },
...
        "/tenant": {
            "get": {
                "description": "this endpoint gets all users under a tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "getAllUsersByTenantID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort_by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort_direction_desc",
                        "name": "sort_direction_desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "audit log fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/model.GenericResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "this endpoint create a new tenent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "createTenant",
                "parameters": [
                    {
                        "description": "tenant request body",
                        "name": "tenantRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tenant.tenantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "tenant created successfully",
                        "schema": {
                            "$ref": "#/definitions/model.GenericResponse"
                        }
                    }
                }
            }
        },
        "/tenant/login": {
            "post": {
                "description": "this endpoint is used to log a user in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "login request body",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tenant.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tenant logged in successfully",
                        "schema": {
                            "$ref": "#/definitions/model.GenericResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.signupRequest": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.updateUserRequest": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
        "model.GenericResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The http response code\n\nRequired: true\nExample: 200",
                    "type": "integer"
                },
                "data": {
                    "description": "The http response data in cases where the request was processed successfully (optional)\n\nExample: {\"id\": \"uuid\", \"name\": \"john doe\"}"
                },
                "error": {
                    "description": "The error message (optional)\n\nExample: cannot process this request at this time (optional)",
                    "type": "string"
                },
                "message": {
                    "description": "The success message (optional)\n\nExample: User has been created successfully (optional)",
                    "type": "string"
                },
                "page": {
                    "description": "Page is the pagination info"
                }
            }
        },
...
        "tenant.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "tenant.tenantRequest": {
            "type": "object",
            "required": [
                "businessName",
                "email",
                "password"
            ],
            "properties": {
                "businessName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}